client_app/requirements.txt
google-adk
streamlit
requests
google-auth